requirements.txt
accelerator==2024.6.11
aiohappyeyeballs==2.3.7
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
bottle==0.12.25
certifi==2024.7.4
charset-normalizer==3.3.2
datasets==2.21.0
dill==0.3.8
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pyarrow==17.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
setproctitle==1.3.3
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
xxhash==3.5.0
yarl==1.9.4
